Clean up field placements
I don't even know if what I did was correct, it just made sense to me as an edit, even though I don't know HTML.
When looking at the contacts page on my computer, it kinda messes up the sections for the New Contacts and Search Contacts. The two fields for Name and Phone Number seem like they should be together on one line, but it seems like it forces them into two lines with a weird indent on the second line, depending on the user's platform I guess. I wish I got a screenshot earlier when the page was up, but I think it would be better to just have them on separate lines. I think the change I made would do that.